Moved ClientError to 'errors' module
