+++ src/actions/attend.ts
@@ -1,6 +1,6 @@
 import { ContextMessageUpdate } from 'telegraf'
 
-import ClientError from '../classes/ClientError'
+import ClientError from '../errors'
 import { getSheetsClient } from '../services/sheetsClient'
 import { updatePlayerCount } from '../services/tableManager'
 import { handleActionError } from '../errors'
