--- src/classes/ClientError.ts
@@ -1,1 +1,0 @@
-export default class ClientError extends Error {}
